Added ES6 support via Babel and switched build system from Gulp 3.9 to Gulp 4.0 alpha, right now this branch is to be concidered broken, until further notice.
