+++ package.json
@@ -1,9 +1,11 @@
 {
   "private": true,
   "devDependencies": {
+    "babel-preset-es2015": "^6.22.0",
+    "babel-register": "^6.23.0",
     "browserify": "^14.0.0",
     "del": "^2.2.2",
-    "gulp": "^3.9.1",
+    "gulp": "github:gulpjs/gulp#4.0",
     "gulp-autoprefixer": "^3.1.0",
     "gulp-clean-css": "^2.3.2",
     "gulp-environments": "^0.1.1",
@@ -19,7 +21,6 @@
     "gulp-uglify": "^1.5.4",
     "gulp-util": "^3.0.7",
     "gulp-watch": "^4.3.9",
-    "run-sequence": "^1.2.2",
     "vinyl-buffer": "^1.0.0",
     "vinyl-ftp": "^0.5.0",
     "vinyl-source-stream": "^1.1.0"
